move `createMotionProxy` into `motion-proxy.ts`
