+++ src/render/dom/motion-minimal.ts
@@ -1,5 +1,5 @@
 import { MeasureLayout } from "../../motion/features/layout/Measure"
-import { createMotionProxy } from "./motion"
+import { createMotionProxy } from "./motion-proxy"
 
 /**
  * @public
