+++ src/render/dom/types.ts
@@ -180,3 +180,5 @@ export type SVGMotionComponents = {
         SVGMotionProps<UnwrapSVGFactoryElement<JSX.IntrinsicElements[K]>>
     >
 }
+
+export type MotionComponents = HTMLMotionComponents & SVGMotionComponents
